+++ packages/components/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Internal
 
 -   `ValidatedCheckboxControl`: Expose the component under private API's ([#71505](https://github.com/WordPress/gutenberg/pull/71505/)).
+-   Expose `ValidatedTextareaControl` via Private APIs ([#71495](https://github.com/WordPress/gutenberg/pull/71495))
 
 ## 30.3.0 (2025-09-03)